separated admin users and classes into separate functions
