+++ src/pages/dashboard/AdminDashboard.tsx
@@ -1,6 +1,3 @@
-import { useState } from "react";
-import { useAllUsers } from "../../hooks/useAllUsers";
-import { useUserClasses, useUserClassStatus } from "../../hooks/useUserClasses";
 import {
   Chart as ChartJS,
   LineElement,
@@ -10,15 +7,7 @@ import {
   Tooltip,
   Legend,
 } from "chart.js";
-import { ClassData, User, UserRole } from "../../api/types/user";
 import "react-datepicker/dist/react-datepicker.css";
-import { useUserActivity } from "../../hooks/useUserActivity";
-import UserSideBar from "../../components/UsersSideBar";
-import UserDetailsPanel from "../../components/UserDetailsPanel";
-import DataDownload from "../../components/DataDownload";
-import { useInstructorClasses } from "../../hooks/useInstructorClasses";
-import { toast } from "sonner";
-import StudentDashboardCard from "../../components/StudentDashboardCard";
 
 ChartJS.register(
   LineElement,
@@ -37,105 +26,7 @@ ChartJS.register(
  * @returns {JSX.Element} The AdminDashboard component.
  */
 export const AdminDashboard = () => {
-  const [selectedUsers, setSelectedUsers] = useState<User[]>([]);
-
-  const [selectedUserId, setSelectedUserId] = useState<string | null>(null);
-  const [selectedClass, setSelectedClass] = useState<ClassData | null>(null);
-  const selectedClassId = selectedClass?.id ?? "all";
-  const selectedClassType =
-    selectedClass?.id === "all"
-      ? "all"
-      : selectedClass?.id === "non-class"
-        ? "non-class"
-        : "class";
-
-  const { users, isLoading, error } = useAllUsers();
-
-  const primaryUser = selectedUsers[0];
-
-  const { classes, loading: userClassesLoading } = useUserClasses(
-    primaryUser?.id
-  );
-
-  const { userActivity, loading: userActivityLoading } = useUserActivity(
-    primaryUser?.id
-  );
-
-  const { classes: instructorClasses, loading: instructorLoading } =
-    useInstructorClasses(primaryUser?.id);
-
-  const { userActivity: selectedActivity, progressData } = useUserActivity(
-    selectedUserId,
-    selectedClassId,
-    selectedClassType
-  );
-
-  const { studentStatus } = useUserClassStatus(
-    selectedUsers[0]?.id || null,
-    selectedClass?.id || null
-  );
-
-  if (error) {
-    toast.error("Error fetching users. Please try again later.");
-    return null;
-  }
-
-  return (
-    <div>
-      {/* User Section */}
-      <div className="flex flex-col md:flex-row gap-6 mb-6">
-        {/* Sidebar */}
-        <UserSideBar
-          users={users}
-          selectedUsers={selectedUsers}
-          loading={isLoading}
-          onSelectUser={(user) => {
-            setSelectedUsers([user]);
-          }}
-          onSetSelectedUsers={setSelectedUsers}
-          setSelectedUserId={setSelectedUserId}
-        />
-
-        {/* Details Panel */}
-        <UserDetailsPanel
-          user={selectedUsers}
-          userClasses={classes}
-          userRole={UserRole.DEV}
-          userActivity={userActivity}
-          instructorClasses={instructorClasses}
-          isLoading={
-            isLoading ||
-            userClassesLoading ||
-            userActivityLoading ||
-            instructorLoading
-          }
-          setSelectedClass={setSelectedClass}
-          isSettingsPanel={false}
-        />
-      </div>
-      <div className="w-full mb-6 text-text">
-        <DataDownload />
-      </div>
-
-      {selectedClass && selectedUserId && (
-        <StudentDashboardCard
-          student={{
-            fullName: `${selectedUsers[0].first_name} ${selectedUsers[0].last_name}`,
-            classTitle: selectedClass.class_title,
-            studentStatus:
-              selectedClassId !== "all" && selectedClassId !== "non-class"
-                ? studentStatus
-                : null,
-            totalAccepted: progressData.totalAccepted,
-            correctSuggestions: progressData.correctSuggestions,
-            percentageCorrect: progressData.percentageCorrect,
-            logs: selectedActivity,
-          }}
-          onClose={() => setSelectedClass(null)}
-        />
-      )}
-    </div>
-  );
+  return <div>{/* User Section */}</div>;
 };
 
 export default AdminDashboard;
